Version 0.7.5
Implemented basic pregeneration of dungeon levels

+++ C-Engine/src/main/java/main/system/auxiliary/secondary/GeometryMaster.java
@@ -113,7 +113,7 @@ public class GeometryMaster {
         return p;
     }
 
-    public static float hyp(int a, int b) {
+    public static float hyp(float a, float b) {
         return (float) Math.sqrt(a * a + b * b);
     }
 }
